refactor(observability): restructure sentry integration
